Fix compiler warnings

+++ benchmark/allocator/allocator.cc
@@ -52,7 +52,7 @@ class AllocatorTestSuite {
   void AllocateAndFreeFixedSize(size_t count, size_t size) {
     Timer t;
     t.Resume();
-    for (int i = 0; i < count; ++i) {
+    for (size_t i = 0; i < count; ++i) {
       Pointer p = alloc_->Allocate(size);
       alloc_->Free(p);
     }
@@ -67,10 +67,10 @@ class AllocatorTestSuite {
     Timer t;
     std::vector<Pointer> cache(count);
     t.Resume();
-    for (int i = 0; i < count; ++i) {
+    for (size_t i = 0; i < count; ++i) {
       cache[i] = alloc_->Allocate(size);
     }
-    for (int i = 0; i < count; ++i) {
+    for (size_t i = 0; i < count; ++i) {
       alloc_->Free(cache[i]);
     }
 #pragma omp barrier
